deploy/requirements.txt
clickhouse-driver==0.2.9
kafka-python==2.2.15
pytz==2025.2
tzlocal==5.3.1
